+++ src/index.tsx
@@ -16,7 +16,7 @@ import Dashboard from "./frames/Dashboard";
 import Login from "./frames/Login";
 import Source from "./models/source";
 import { FirebaseUser } from "./models/user";
-import CreateOrRoute from "./pages/createpage/CreateOrRoute";
+import CreateOrRoute from "./pages/createpage/Route";
 import IntegrationPage from "./pages/integration/StateIntegrationPage";
 import LoginPage from "./pages/LoginPage";
 import LogsPage from "./pages/logspage/ConvoPage";
