+++ src/pages/createpage/Route.test.tsx
@@ -9,7 +9,7 @@ import Source from "../../models/source";
 import User from "../../models/user";
 import SourceService from "../../services/source";
 import { dummySources } from "../../utils/test";
-import { CreateOrRoute } from "./CreateOrRoute";
+import { Route } from "./Route";
 
 chai.use(sinonChai);
 const expect = chai.expect;
@@ -45,15 +45,15 @@ describe("Route", function () {
         it("Tests default route when no parameters in place.", function () {
             const location = { query: {} };
             // It calls it on mount.
-            shallow(<CreateOrRoute currentUser={user} goTo={goTo} location={location} />);
+            shallow(<Route currentUser={user} goTo={goTo} location={location} />);
             expect(goTo).to.have.been.calledWith("/skills");
         });
 
         it("Tests default route when when only ID exists.", function () {
             const location = { query: { id: returnSources[0].id } };
             // It calls it on mount.
-            const wrapper = shallow(<CreateOrRoute currentUser={user} goTo={goTo} location={location} />);
-            const instance = wrapper.instance() as CreateOrRoute;
+            const wrapper = shallow(<Route currentUser={user} goTo={goTo} location={location} />);
+            const instance = wrapper.instance() as Route;
 
             return (instance.cancelables[0] as Bluebird<any>).finally(function () {
                 expect(goTo).to.have.been.calledWith("/notFound");
@@ -63,8 +63,8 @@ describe("Route", function () {
         it("Tests default route when when only key exists exists.", function () {
             const location = { query: { key: returnSources[0].secretKey } };
             // It calls it on mount.
-            const wrapper = shallow(<CreateOrRoute currentUser={user} goTo={goTo} location={location} />);
-            const instance = wrapper.instance() as CreateOrRoute;
+            const wrapper = shallow(<Route currentUser={user} goTo={goTo} location={location} />);
+            const instance = wrapper.instance() as Route;
 
             return (instance.cancelables[0] as Bluebird<any>).finally(function () {
                 expect(goTo).to.have.been.calledWith("/notFound");
@@ -74,8 +74,8 @@ describe("Route", function () {
         it("Tests it routes to the skill when found.", function () {
             const location = { query: { id: returnSources[3].id, key: returnSources[3].secretKey } };
             // It calls it on mount.
-            const wrapper = shallow(<CreateOrRoute currentUser={user} goTo={goTo} location={location} />);
-            const instance = wrapper.instance() as CreateOrRoute;
+            const wrapper = shallow(<Route currentUser={user} goTo={goTo} location={location} />);
+            const instance = wrapper.instance() as Route;
 
             return (instance.cancelables[0] as Bluebird<any>).finally(function () {
                 expect(goTo).to.have.been.calledWith("/skills/" + returnSources[3].id);
@@ -85,8 +85,8 @@ describe("Route", function () {
         it("Tests it routes to the default when both parameters exist and not found.", function () {
             const location = { query: { id: returnSources[3].id, key: "gibberish" } };
             // It calls it on mount.
-            const wrapper = shallow(<CreateOrRoute currentUser={user} goTo={goTo} location={location} />);
-            const instance = wrapper.instance() as CreateOrRoute;
+            const wrapper = shallow(<Route currentUser={user} goTo={goTo} location={location} />);
+            const instance = wrapper.instance() as Route;
 
             return (instance.cancelables[0] as Bluebird<any>).finally(function () {
                 expect(goTo).to.have.been.calledWith("/notFound");
